refactor: move worker to helpers
[no ci]
